requirements.txt
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cycler==0.12.1
dm-tree==0.1.8
Farama-Notifications==0.0.4
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
gymnasium==0.28.1
hyperopt==0.2.7
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
jax-jumpy==1.0.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
py4j==0.10.9.7
pyarrow==16.1.0
Pygments==2.18.0
pymap3d==3.1.0
pyparsing==3.1.2
pytest==8.2.2 
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.10.0
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
Rtree==1.2.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-image==0.22.0
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
# -e git+ssh://git@github.com/decisionLabLtd/dstl-mast.git@9850933197009afffde95107aa33755e8c785613#egg=stonesoup&subdirectory=StoneSoup
sympy==1.12.1
tensorboardX==2.6.2.2
tifffile==2024.5.22
torch==2.3.1
tqdm==4.66.4
triton==2.3.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
utm==0.7.0
zipp==3.18.2
